changed to scripted pipeline

+++ Jenkinsfile
@@ -1,52 +1,39 @@
-pipeline {
-    agent none
-    options {
-        skipStagesAfterUnstable()
+node {
+    stage('Build') {
+        withDockerContainer('python:2-alpine') {
+            sh 'python -m py_compile sources/add2vals.py sources/calc.py'
+            stash includes: 'sources/*.py*', name: 'compiled-results'
+        }
     }
-    stages {
-        stage('Build') {
-            agent {
-                docker {
-                    image 'python:2-alpine'
-                }
-            }
-            steps {
-                sh 'python -m py_compile sources/add2vals.py sources/calc.py'
-                stash(name: 'compiled-results', includes: 'sources/*.py*')
-            }
-        }
+    try {
         stage('Test') {
-            agent {
-                docker {
-                    image 'qnib/pytest'
-                }
-            }
-            steps {
+            withDockerContainer('qnib/pytest') {
                 sh 'py.test --junit-xml test-reports/results.xml sources/test_calc.py'
-            }
-            post {
-                always {
-                    junit 'test-reports/results.xml'
-                }
+                stash includes: 'sources/*.py*', name: 'compiled-results'
             }
         }
-        stage('Deliver') { 
-            agent any
-            environment { 
-                VOLUME = '$(pwd)/sources:/src'
-                IMAGE = 'cdrx/pyinstaller-linux:python2'
-            }
-            steps {
-                dir(path: env.BUILD_ID) { 
+    } catch (e) {
+        echo 'Terjadi kesalahan'
+        throw e
+    } finally {
+        junit 'test-reports/results.xml'   
+    }
+    stage('Deliver') { 
+        try {
+            withEnv(['VOLUME = \'$(pwd)/sources:/src\'', 'IMAGE = \'cdrx/pyinstaller-linux:python2\'']) {
+                dir('env.BUILD_ID') {
                     unstash(name: 'compiled-results') 
-                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'pyinstaller -F add2vals.py'" 
+                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'pyinstaller -F add2vals.py'"
                 }
             }
-            post {
-                success {
-                    archiveArtifacts "${env.BUILD_ID}/sources/dist/add2vals" 
-                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
-                }
+        } catch (e) {
+            echo 'Terjadi kesalahan'
+            throw e
+        } finally {
+            def currentResult = currentBuild.result
+            if (currentResult == 'SUCCESS') {
+                archiveArtifacts "${env.BUILD_ID}/sources/dist/add2vals" 
+                sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
             }
         }
     }
